pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cuda-data-operators"
version = "0.1.0"
description = "GPU-native data operators for SQL-like operations"
requires-python = ">=3.11,<3.14"
dependencies = [
    "numpy>=1.24.0",
    "cupy-cuda12x>=12.0.0; sys_platform == 'linux'",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["cuda_dataops"]

[project.scripts]
test-ops = "test_runner:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]  # line length handled separately

[tool.ruff.lint.isort]
known-first-party = ["cuda_data_operators"]
